Working on converging to using utils.py for actual shared code
Fixes #23
