+++ repo_activity.py
@@ -23,29 +23,7 @@ import utils
 # This is the number of retries, otherwise, just report the problem in the output and move along
 MAX_RETRIES = 5
 # Roughly the number of github queries per loop.  Guessing bigger is better
-RATE_PER_LOOP = 6
-
-
-def _create_char_spinner():
-    """
-    Creates a generator yielding a char based spinner.
-    """
-    while True:
-        for char in "|/-\\":
-            yield char
-
-
-_spinner = _create_char_spinner()
-
-
-def spinner(label=""):
-    """
-    Prints label with a spinner.
-    When called repeatedly from inside a loop this prints
-    a one line CLI spinner.
-    """
-    sys.stderr.write("\r%s %s" % (label, next(_spinner)))
-    sys.stderr.flush()
+RATE_PER_LOOP = 20
 
 
 def parse_args():
@@ -95,29 +73,6 @@ def parse_args():
     if args.token is None:
         args.token = getpass("Please enter your GitHub token: ")
     return args
-
-
-def check_rate_remain(gh_sess, loopsize, update=False):
-    """
-    Given the session, and the size of the rate eaten by the loop,
-    and if not enough remains, sleep until it is.
-    :param gh_sess: The github session
-    :param loopsize: The amount of rate eaten by a run through things
-    :param update: Should we print a progress element to stderr
-    """
-    # TODO: Look at making the naptime show that you're still making progress
-    while gh_sess.rate_limit()["resources"]["core"]["remaining"] < loopsize:
-        # Uh oh.
-        if update:
-            sleep(5)
-            spinner()
-        else:
-            # calculate how long to sleep, sleep that long.
-            refreshtime = datetime.fromtimestamp(gh_sess.rate_limit()["resources"]["core"]["reset"])
-            now = datetime.now()
-            naptime = (refreshtime - now).seconds
-            print(f"Sleeping for {naptime} seconds", file=sys.stderr)
-            sleep(naptime)
 
 
 def repo_activity(gh_sess, org, repo):  # pylint: disable=too-many-branches
@@ -255,16 +210,19 @@ def main():
                             file=sys.stderr,
                         )
                     done = True
-                check_rate_remain(gh_sess, RATE_PER_LOOP, args.info)
+                utils.check_rate_remain(gh_sess, RATE_PER_LOOP, args.info)
 
                 if result == 202:
                     sleep(10)
                 if args.info:
-                    spinner()
+                    utils.spinner()
         else:
             mini_repo_activity(gh_sess, org, repo)
             if args.info:
-                spinner()
+                utils.spinner()
+
+    if args.info:
+        print(file=sys.stderr)
 
 
 if __name__ == "__main__":
